Revert "[Fix] Change 'long int' to 'unsigned int'"
BUG found in:
  FILE: random.c
  FUNCTION: random_in_range_exclusive

  Overflow in arithmetic multiplication operation.

The rand() function from the <stdlib.h> library generates a number in
the range [0, RAND_MAX]. In the GNU library, RAND_MAX is 2147483647,
which is the largest signed integer representable in 32 bits. The issue
raises when executing: rand() * length; being 'length' declared as
'unsigned int'. GCC compiles gives an overflow warning.

According to the C standard ISO/IEC 9899:201x (2011) Section 6.3.1.1-2,
both, the number generated by 'rand()' and 'length' are promoted to
'int' since it can represent all their possible values.
This, leads to GCC to warn us because the operation might cause an
overflow which is considered 'undefined behaviour'.

Additional reference:
http://stackoverflow.com/questions/15656437/int-promotion-is-the-following-well-defined

This reverts commit 9967e88a1bb5ee6e455d3a51ec2ddd47619bb30f.