[*] Improve the seed generation
random.c:initialize_random_sequence

Problem: the same sequence of random numbers will be generated if the
function is called more than once within a second, given that the
function 'time' returns the number of seconds since the 'epoch time'.

Solution: use the function 'gettimeofday' to use both, seconds and
microseconds; so that, now, the function would have to be called several
times within a microsecond to observe the problem. We assume this is not
possible with current hardware.