[*] Use function-like macros for error reporting
[+] report.h:ERROR
[+] report.h:ERROR_VERBOSE
[-] report.c Not needed anymore

Why macros? Although macros should be used sparingly and carefully,
there are cases where they are useful. In this case, we use them to hide
the automatic insertion of the file (__FILE__) and function
(__func__) names for error reporting, which cannot be done by using
the functions: 'error' and 'error_verbose' without explicitly indicating
the names.

References:
https://gcc.gnu.org/onlinedocs/gcc/Function-Names.html
http://stackoverflow.com/questions/1644868/c-define-macro-for-debug-printing