[*] Enforce implicit binary-encoded individuals
[-] genome.h:
    struct Encoding: *nucleotides,
    struct Encoding: nucleotides_length,
    min_nucleotide_value,
    max_nucleotide_value,
[-] genome.c:
    min_nucleotide_value
    max_nucleotide_value
[+] random.h:randomize_bins
[+] random.c:randomize_bins

The attributes *nucleotides and nucleotides_length, were the means by
which we intended to provide the flexibility to specify different
alphabets to encode individuals. As an example, if we choose
'*nucleotides' to be: {0, 1} it would be a binary-based encoding,
whereas if we choose it to be: {0, 1, 2, ..., N} it would correspond to
an integer-based encoding. Such flexibility is unnecessary if we
restrict DARWIN to use only the binary format.

DARWIN will be designed such that the functions to manipulate genes can
only generate zeroes and ones; thus, implicitly enforcing the encoding.