[+] Add support for multiple-valued alleles
[+] genome.h:
    MIN_UNITS_PER_GENE
    struct Encoding: units_per_gene
    dna_byte_size
[+] genome.c:dna_byte_size

[*] genome.h. Use better variable/macro names
[*] genome.c. Use better variable/macro names

To represent multiple-valued alleles in a binary format, we must be able
to express genes having more than one unit (or nucleotide). That is, if
we have genes of size N, each gene can represent 2^N allele values. For
example, say we have an individual with DNA:
    7-1-4-2-2
its equivalent binary representation would be:
    (111)-(001)-(100)-(010)-(010)
each gene having 3 units or nucleotides