[*] Move <assert.h> from "report.h" to "random.h"
"report.h" does not use the 'assert' macro. If it is not required, then
it should not be included. However, "random.h" is the lowest
abstraction level where 'assert' is required.

Note: Although the first time 'assert' is called is in the invocation to
'create_encoding' in genome.c, I prefered to hide the details in the
lower levels.