[-] Population is initialized at its maximum size
Previously we could initialize it with size '4' and a max. size of '10'.
That might be the appropiate approach for an implementation in which the
size varies through generations. However, we take the simplest approach
for this project: constant size after initialization.

One subtle, but important, detail is that in this simplest approach we
maximize the number of individuals generated randomly. Hence, providing
more 'exploration' at the start of the algorithm; resembling a similar
idea used in 'Simulated Annealing'.