[*] single_point_crossover: better args and return
operators.{h,c}: single_point_crossover

    what?

Instead of passing the 'offspring' pointers, 'son' and 'daugher', as
arguments: create them inside, do the crossover, and return them.

    why?

Creating the offspring outside the function would be equivalent to say:
"take out of the function part of its functionality". In this case, it
is simply wrong.