Add darwin's configuration files support
darwin/data
    * Autogeneration scripts
    * Configuration files.

darwin/src/experiment/io/in
    * Parsers

The support introduced will enable the user to tell darwin which problem
to solve. To that end, the autogeneration scripts allow the user to
easily generate pseudo-random instances for certain problems (i.e
Subsetsum or Hamiltonian Path) that otherwise would be tedious to write
by hand. Apart from these scripts other pre-defined configuration files
will be included in the future; in order to test specific instances as
well. This way, the user will just need to provide the absolute path to
test one of the problem instances.