Add X-Macro mechanism
In the C programming language there are several ways in which we can
simulate polymorphism. One way would be using 'void *' as the type in
which to encapsulate the data (char, int, double, struct, union, ...),
but at the expense of loosing type safety. Furthermore, we still need to
define all the specific functions for each of the data types we want to
support; leading us to have to add additional code to handle type
checking, conversion from (to) 'void *' to (from) the specific data
types. The consequences are that code generally becomes harder to
maintain, harder to understand, harder to use, and harder to extend.

Instead, we can take advantage of the C preprocessor and define macros
that can 'do the job' for us. Although, macros always must be used
prudently or avoided if possible, there are certain cases for which they
are a more convenient tool. X-Macro is just a particular technique to
use macros (https://en.wikipedia.org/wiki/X_Macro).

The result of using this technique is that now would be easier to
integrate new problem sets to darwin, since we simply need to
incorporate the associated code for the internals of the problem set.